# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shop API",
    "version": "1.0.0",
    "description": "Практики: auth (JWT + refresh), пользователи, товары, RBAC."
  },
  "servers": [{ "url": "/", "description": "Текущий хост (например http://localhost:3000)" }],
  "tags": [
    { "name": "Auth" },
    { "name": "Users", "description": "Только admin" },
    { "name": "Products" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token из POST /api/auth/login"
      }
    },
    "schemas": {
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "seller", "admin"] },
          "blocked": { "type": "boolean" }
        }
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Регистрация",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "first_name", "last_name", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создан",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUser" } } }
          },
          "400": { "description": "Неверное тело", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email занят", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Вход",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токены",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tokens" } } }
          },
          "401": { "description": "Неверные данные", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Обновление пары токенов",
        "parameters": [
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Refresh token из login"
          }
        ],
        "responses": {
          "200": {
            "description": "Новые токены",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tokens" } } }
          },
          "400": { "description": "Нет заголовка" },
          "401": { "description": "Невалидный refresh" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Текущий пользователь",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUser" } } }
          },
          "401": { "description": "Нет или просрочен access token" }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Список пользователей",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PublicUser" } } }
            }
          },
          "403": { "description": "Не admin" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Пользователь по id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUser" } } } },
          "404": { "description": "Не найден" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Изменить пользователя (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "email": { "type": "string" },
                  "role": { "type": "string", "enum": ["user", "seller", "admin"] },
                  "blocked": { "type": "boolean" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUser" } } } },
          "404": {},
          "409": { "description": "Email занят" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Заблокировать пользователя",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Заблокирован" },
          "400": { "description": "Нельзя заблокировать себя" }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Список товаров",
        "security": [{ "bearerAuth": [] }],
        "description": "Роли: user, seller, admin",
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } }
            }
          },
          "403": {}
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Создать товар",
        "security": [{ "bearerAuth": [] }],
        "description": "Роли: seller, admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "category", "description", "price"],
                "properties": {
                  "title": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "201": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } }, "400": {} }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "404": {}
        }
      },
      "put": {
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "description": "Роли: seller, admin",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } }, "404": {} }
      },
      "delete": {
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "description": "seller и admin могут удалять товар",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Удалено" }, "404": {} }
      }
    }
  }
}